Created executable, update requirements, clean up file organization
