+++ requirements.txt
@@ -1,4 +1,4 @@
 numpy==1.21.4
-opencv-python==4.5.4.58
+opencv-python==4.5.4.60
 Pillow==8.4.0
 pytesseract==0.3.8
